+++ vite.config.ts
@@ -1,8 +1,16 @@
 import { defineConfig } from "vite";
+import inject from "@rollup/plugin-inject";
 import react from "@vitejs/plugin-react";
 import tsconfigPaths from "vite-tsconfig-paths";
 
 // https://vitejs.dev/config/
 export default defineConfig({
-	plugins: [tsconfigPaths(), react()],
+	plugins: [
+		inject({
+			$: "jquery",
+			jQuery: "jquery",
+		}),
+		tsconfigPaths(),
+		react(),
+	],
 });